Useless module `util` Removed
