+++ cmd/nem/main.go
@@ -15,9 +15,10 @@ import (
 
 	"math"
 
+	"encoding/hex"
+
 	"github.com/nem-toolchain/nem-toolchain/pkg/core"
 	"github.com/nem-toolchain/nem-toolchain/pkg/keypair"
-	"github.com/nem-toolchain/nem-toolchain/pkg/util"
 	"github.com/nem-toolchain/nem-toolchain/pkg/vanity"
 	"github.com/urfave/cli"
 )
@@ -102,7 +103,7 @@ func generateAction(c *cli.Context) error {
 
 	num := c.Uint("number")
 	for i := uint(0); i < num; i++ {
-		util.PrintAccountDetails(ch, keypair.Gen())
+		printAccountDetails(ch, keypair.Gen())
 		fmt.Println("----")
 	}
 
@@ -148,7 +149,16 @@ func vanityAction(c *cli.Context) error {
 				fmt.Print(".")
 			}
 		}()
-		rate := util.CPUKeyPairsInSeconds()
+		rate := float64(0)
+		res := make(chan int, runtime.NumCPU())
+		for i := 0; i < cap(res); i++ {
+			go func(res chan<- int) {
+				res <- keypair.CountKeyPairs(3200)
+			}(res)
+		}
+		for i := 0; i < cap(res); i++ {
+			rate += float64(<-res) / 3.2
+		}
 		ticker.Stop()
 		fmt.Printf(" %v accounts/sec\n", math.Trunc(rate))
 
@@ -157,9 +167,9 @@ func vanityAction(c *cli.Context) error {
 			fmt.Printf("Specified search complexity: %v\n", math.Trunc(1.0/pbty))
 		}
 		fmt.Printf("Estimate search times: %v (50%%), %v (80%%), %v (99.9%%)\n",
-			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.5)/rate),
-			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.8)/rate),
-			util.TimeInSeconds(util.NumberOfKeyPairs(pbty, 0.99)/rate))
+			timeInSeconds(vanity.NumberOfKeyPairs(pbty, 0.5)/rate),
+			timeInSeconds(vanity.NumberOfKeyPairs(pbty, 0.8)/rate),
+			timeInSeconds(vanity.NumberOfKeyPairs(pbty, 0.99)/rate))
 		fmt.Println("----")
 	}
 
@@ -172,7 +182,7 @@ func vanityAction(c *cli.Context) error {
 		if i != 0 {
 			fmt.Println("----")
 		}
-		util.PrintAccountDetails(ch, <-rs)
+		printAccountDetails(ch, <-rs)
 		go vanity.StartSearch(ch, sel, rs)
 	}
 
@@ -192,4 +202,20 @@ func chainGlobalOption(c *cli.Context) (core.Chain, error) {
 		return ch, fmt.Errorf("unknown chain '%v'", c.GlobalString("chain"))
 	}
 	return ch, nil
+}
+
+// Format estimated time
+func timeInSeconds(val float64) string {
+	val = 1e9 * math.Trunc(val)
+	if val >= math.MaxInt64 || math.IsInf(val, 0) {
+		return "Inf"
+	}
+	return time.Duration(val).String()
+}
+
+// Pretty print account details
+func printAccountDetails(chain core.Chain, pair keypair.KeyPair) {
+	fmt.Println("Address:", pair.Address(chain).PrettyString())
+	fmt.Println("Public key:", hex.EncodeToString(pair.Public))
+	fmt.Println("Private key:", hex.EncodeToString(pair.Private))
 }
