+++ pkg/vanity/probability.go
@@ -30,6 +30,12 @@ func Probability(sel Selector) float64 {
 	return res
 }
 
+// Calculate amount of keypairs to be generated
+// to find account with probability `pbty`
+func NumberOfKeyPairs(pbty, prec float64) float64 {
+	return math.Log(1-prec) / math.Log(1-pbty)
+}
+
 func (rule searchRule) probability() float64 {
 	res := float64(1)
 	if rule.noDigitSelector != nil && rule.prefixSelector != nil {
